Text- ja PicPresentation testien poisto
