+++ project/GUI/remu.kv
@@ -82,7 +82,7 @@ RemuSM:
             Button:
                 size_hint: 0.3,0.3
                 text: "Go to presentation"
-                on_press: root.pic_presentation_chosen()
+                on_press: root.init_presentation()
                 on_press: app.root.change_screen_to("presentation_layout")
             Button:
                 size_hint:0.3,0.3
